Changes to support Ubuntu 24.04
Remove python3-binutils and replace with python3-setuptools. Update the
localcoords setup script for the change. Also update the CI configs to
use Ubuntu 24.04/pythony 3.12

Note: the localcoords script is trying to do a systemwide install of a
C++ module for use in python. This is probably not what's best for
OpenUxAS users.